Create Alternate matrix
Alternate Matrix Addition
You are given elements of 2D matrix of size of N*N. The task is to print the sum of elements of arrays such that the individual element is alternative to each other in the same array and alternate with other array elements as well. Obviously all the elements of matrix cannot be covered in single step since they should be alternative, thus you need one more step to cover the rest of the elements but again keeping in mind of alternative rule described above. Note: Start from the first element of the matrix and the matrix will always be squared.
Input
There will be two lines in input, the first line will contain N. Second line will contain the elements in matrix form.

Your Task:
You will be given a function alternate_Matrix_Sum() to complete it. The function will contain the input 2D array and N.
Output
First line should print sum of A[0][0], A[0][2], A[1][1], A[2][0], A[2][2]

Second line should print sum of A[0][1], A[1][0], A[1][2], A[2][1]